refactor: ref episode
