+++ src/view/__tests__/index.test.js
@@ -4,9 +4,9 @@ jest.mock('react-dom/client', () => ({
   createRoot: jest.fn().mockImplementation(() => ({ render: jest.fn() })),
 }));
 
-jest.mock('../containers/App/App.tsx', () => () => 'MyComponent');
+jest.mock('../containers/App/App', () => () => 'MyComponent');
 
-import '../index.tsx';
+import '../index';
 
 describe('root render unit test', () => {
   afterAll(jest.clearAllMocks);
