+++ src/view/components/SearchSkeleton/__tests__/SearchSkeleton.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import SearchSkeleton from '../SearchSkeleton.tsx';
+import SearchSkeleton from '../SearchSkeleton';
 
 jest.mock('@containers/search/components/Channel/Channel.jsx', () => () => (
   <div id="channel-mock">Hola, soy un item de channel</div>
