--- src/view/containers/episode/components/index.js
@@ -1,1 +1,0 @@
-export { default } from './EpisodeDetail';
