+++ src/view/containers/search/Search.jsx
@@ -4,7 +4,7 @@ import styles from './styles.module.scss';
 import SearchBox from './components/SearchBox/SearchBox.jsx';
 import { channels as channelsQuery } from '../../../queries/podcasts.js';
 import { useIsFetching, useQuery } from '@tanstack/react-query';
-import SearchSkeleton from '@components/SearchSkeleton/SearchSkeleton.tsx';
+import SearchSkeleton from '@components/SearchSkeleton/SearchSkeleton';
 
 const Search = () => {
   const { data: channels } = useQuery(channelsQuery());
